requirements.txt
watchdog
azure-storage-queue
boto3
localstack
awscli-local